requirements.txt
# 기존 패키지들
google-generativeai>=0.3.0
biopython>=1.81
pandas>=1.5.0
plotly>=5.0.0
yattag>=1.14.0

# 보안을 위한 환경변수 관리
python-dotenv>=1.0.0

# 추가 유용한 패키지들
requests>=2.28.0
numpy>=1.24.0